Added an argument label to the colorbar
I seriously didn't expect that I will add something to this repo this quickly!

You see, I was reading the draft of this project. There, I have included some of the results in the frequency domain with the color bar. However, the values attached to the color bar does not make it clear that I am using the dB scale. Therefore, I went ahead and searched if it was possible to add a label to the color bar, and here we are :)