FIXING A CRITICAL BIG!!!
It seems that adding a feature before testing is not exactly a great idea :)

The issue is that after adding the argument `cbar_label`, I did not fix `lazy_imshow` to account for it, which caused a mess. Good thing I have tackled this relatively quickly (goes to show how much I use the `lazy_imshow`).